requirements.txt
streamlit==1.32.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0
matplotlib==3.8.2
groq==0.3.0
httpx==0.26.0
asyncio==3.4.3
gunicorn